api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
ultralytics>=8.0.0
huggingface_hub>=0.23.0
